# docs/development/api/openapi.json
{
  "components": {
    "schemas": {
      "ErrorBody": {
        "description": "Standard error payload returned by every failed request.\n\nThe envelope is uniform across every error the server emits, including\nrouting fallbacks and extractor rejections (`API-039`).",
        "properties": {
          "_links": {
            "$ref": "#/components/schemas/SelfLinks",
            "description": "Link to the request that produced the error."
          },
          "error": {
            "description": "Human-readable description of the error.",
            "example": "An error message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "_links"
        ],
        "type": "object"
      },
...
      "Link": {
        "description": "A single HAL link.",
        "properties": {
          "href": {
            "description": "Target URI reference of the link.",
            "type": "string"
          },
          "templated": {
            "description": "Whether `href` is a URI template, per RFC 6570.",
            "example": true,
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "required": [
          "href"
        ],
        "type": "object"
      },
...
      "PostUploadRequest": {
        "description": "Payload initializing an upload session.",
        "properties": {
          "content_type": {
            "description": "Declared media type of the file content.",
            "example": "video/mp4",
            "type": "string"
          },
          "file_name": {
            "description": "Name of the file being uploaded.",
            "example": "clip.mp4",
            "minLength": 1,
            "type": "string"
          },
          "file_size": {
            "description": "Total size of the file being uploaded, in bytes.",
            "example": 1048576,
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "md5": {
            "description": "MD5 digest of the complete file, as a 32-character lowercase\nhexadecimal string.",
            "example": "d41d8cd98f00b204e9800998ecf8427e",
            "maxLength": 32,
            "minLength": 32,
            "type": "string"
          }
        },
        "required": [
          "content_type",
          "file_name",
          "file_size",
          "md5"
        ],
        "type": "object"
      },
      "PostUploadResponse": {
        "description": "Upload session created by a successful initialization.",
        "properties": {
          "_links": {
            "$ref": "#/components/schemas/UploadSessionLinks",
            "description": "Links to the created upload session."
          },
          "chunk_size": {
            "description": "Fixed size of every chunk except the last, in bytes.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "expires_at": {
            "description": "Date and time at which the upload session expires, as\n`YYYY-MM-DDTHH:MM:SS`.",
            "example": "2026-01-01T12:00:00",
            "type": "string"
          },
          "upload_id": {
            "$ref": "#/components/schemas/i64",
            "description": "Unique identifier of the upload session."
          }
        },
        "required": [
          "chunk_size",
          "expires_at",
          "_links",
          "upload_id"
        ],
        "type": "object"
      },
...
      "SelfLinks": {
        "description": "The `self` link of a HAL representation.\n\nEvery resource representation carries the URI it was served from, so clients\nnever construct or hardcode it themselves (`API-032`, `API-037`).",
        "properties": {
          "self": {
            "$ref": "#/components/schemas/Link",
            "description": "Link to the resource itself."
          }
        },
        "required": [
          "self"
        ],
        "type": "object"
      },
      "UploadSessionLinks": {
        "description": "HAL links exposed by an upload-session representation.\n\nThe `self` link addresses the session itself; the templated `chunk` link\nlets clients discover where to store a chunk instead of building the URL\nfrom the identifiers (`API-037`); the `complete` link lets clients discover\nhow to finish the session.",
        "properties": {
          "chunk": {
            "$ref": "#/components/schemas/Link",
            "description": "URI template for storing one chunk of the upload session."
          },
          "complete": {
            "$ref": "#/components/schemas/Link",
            "description": "Link completing the upload session."
          },
          "self": {
            "$ref": "#/components/schemas/Link",
            "description": "Link to the upload-session resource itself."
          }
        },
        "required": [
          "chunk",
          "complete",
          "self"
        ],
        "type": "object"
      },
...
      "i64": {
        "format": "int64",
        "type": "integer"
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API of the Mnemorium service",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Mnemorium API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/asset/upload": {
      "post": {
        "description": "Validates the declared content type, file name, size and MD5 digest, creates\nthe upload session and its staging file, and returns the session identifier,\nthe fixed chunk size and the expiry instant.",
        "operationId": "post_upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/PostUploadResponse"
                }
              }
            },
            "description": "Upload session created",
            "headers": {
              "Location": {
                "description": "Canonical URI of the created upload session",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid payload"
          },
          "401": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "413": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The declared file size exceeds the maximum allowed size"
          },
          "415": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The declared content type is not a supported media type"
          },
          "500": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Initialize an upload session",
        "tags": [
          "asset"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://0.0.0.0:4080/api/v1"
    }
  ],
...
}